Advanced R Markdown Word Format

Format for converting from R Markdown to an MS Word document. The function comes also with improved output options.

rdocx_document(
  base_format = "rmarkdown::word_document",
  tables = list(),
  plots = list(),
  lists = list(),
  mapstyles = list(),
  ...
)

Arguments

base_format
a scalar character, format to be used as a base document for officedown. default to word_document but can also be word_document2 from bookdown

tables
a list that can contain few items to style tables and table captions. Missing items will be replaced by default values. Possible items are the following:

style: the Word stylename to use for tables.

layout: 'autofit' or 'fixed' algorithm. See table_layout.

width: value of the preferred width of the table in percent (base 1).

caption; caption options, i.e.:

style: Word stylename to use for table captions.

pre: prefix for numbering chunk (default to "Table ").

sep: suffix for numbering chunk (default to ": ").

conditional: a list of named logical values:

first_row and last_row: apply or remove formatting from the first or last row in the table

first_column and last_column: apply or remove formatting from the first or last column in the table

no_hband and no_vband: don't display odd and even rows or columns with alternating shading for ease of reading.

Default value is (in R format):

list(
   style = "Table", layout = "autofit", width = 1,
   caption = list(
     style = "Table Caption", pre = "Table ", sep = ": "),
   conditional = list(
     first_row = TRUE, first_column = FALSE, last_row = FALSE,
     last_column = FALSE, no_hband = FALSE, no_vband = TRUE
   )
)

Default value is (in YAML format):

style: Table
layout: autofit
width: 1.0
caption:
  style: Table Caption
  pre: 'Table '
  sep: ': '
conditional:
  first_row: true
  first_column: false
  last_row: false
  last_column: false
  no_hband: false
  no_vband: true

plots
a list that can contain few items to style figures and figure captions. Missing items will be replaced by default values. Possible items are the following:

style: the Word stylename to use for plots.

align: alignment of figures in the output document (possible values are 'left', 'right' and 'center').

caption; caption options, i.e.:

style: Word stylename to use for figure captions.

pre: prefix for numbering chunk (default to "Figure ").

sep: suffix for numbering chunk (default to ": ").

Default value is (in R format):

list(
  style = "Normal", align = "center",
  caption = list(
    style = "Image Caption",
    pre = "Figure ",
    sep = ": "
   )
 )

Default value is (in YAML format):

style: Normal
align: center
caption:
  style: Image Caption
  pre: 'Figure '
  sep: ': '

lists
a list containing two named items ol.style and ul.style, values are the stylenames to be used to replace the style of ordered and unordered lists created by pandoc. If NULL, no replacement is made.

Default value is list(ol.style = NULL, ul.style = NULL):

ol.style: null
ul.style: null

mapstyles
a named list of style to be replaced in the generated document. list("Normal" = c("Author", "Date")) will result in a document where all paragraphs styled with stylename "Date" and "Author" will be then styled with stylename "Normal".

...
arguments used by word_document

Value

R Markdown output format to pass to render

Finding stylenames

You can access them in the Word template used. Function styles_info() can let you read these styles.

You need officer to read the stylenames (to get information from a specific "reference_docx", change ref_docx_default in the example below.

library(officer)
docx_file &lt;- system.file(package = "officer", "template", "template.docx")
doc &lt;- read_docx(docx_file)

To read paragraph stylenames:

styles_info(doc, type = "paragraph")

To read table stylenames:

styles_info(doc, type = "table")

To read list stylenames:

styles_info(doc, type = "numbering")

R Markdown yaml

The following demonstrates how to pass arguments in the R Markdown yaml:

---
output:
  officedown::rdocx_document:
    reference_docx: pandoc_template.docx
    tables:
      style: Table
      layout: autofit
      width: 1.0
      caption:
        style: Table Caption
        pre: 'Table '
        sep: ': '
      conditional:
        first_row: true
        first_column: false
        last_row: false
        last_column: false
        no_hband: false
        no_vband: true
    plots:
      style: Normal
      align: center
      caption:
        style: Image Caption
        pre: 'Figure '
        sep: ': '
    lists:
      ol.style: null
      ul.style: null
    mapstyles:
      Normal: ['First Paragraph', 'Author', 'Date']
---

Examples

library(rmarkdown)

# official template -----
skeleton <- system.file(package = "officedown",
  "rmarkdown/templates/word/skeleton/skeleton.Rmd")
rmd_file <- tempfile(fileext = ".Rmd")
file.copy(skeleton, to = rmd_file)

#> [1] TRUE

docx_file_1 <- tempfile(fileext = ".docx")
render(rmd_file, output_file = docx_file_1, quiet = TRUE)

# bookdown example -----

# All above is only to make sure we do not write in your wd
bookdown_loc <- system.file(package = "officedown", "examples/bookdown")
temp_dir <- tempfile()
dir.create(temp_dir, showWarnings = FALSE, recursive = TRUE)
file.copy(
  from = list.files(bookdown_loc, full.names = TRUE),
  to = temp_dir,
  overwrite = TRUE, recursive = TRUE)

#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE

render_site(
  input = temp_dir, encoding = 'UTF-8',
  envir = new.env(), quiet = TRUE)

#> 
#> 
#> processing file: bookdown.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |...                                                                   |   4%
#>    inline R code fragments
#> 
#>   |                                                                              |......                                                                |   8%
#> label: unnamed-chunk-1 (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |........                                                              |  12%
#>    inline R code fragments
#> 
#>   |                                                                              |...........                                                           |  16%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ eval: logi FALSE
#> 
#>   |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................                                                     |  24%
#> label: unnamed-chunk-3
#>   |                                                                              |....................                                                  |  28%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................                                                |  32%
#> label: unnamed-chunk-4
#>   |                                                                              |.........................                                             |  36%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: unnamed-chunk-5 (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |...............................                                       |  44%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................                                    |  48%
#> label: unnamed-chunk-6 (with options) 
#> List of 3
#>  $ tab.cap     : chr "mtcars"
#>  $ tab.id      : chr "mtcars"
#>  $ first_column: logi TRUE
#> 
#>   |                                                                              |....................................                                  |  52%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......................................                               |  56%
#> label: unnamed-chunk-7 (with options) 
#> List of 4
#>  $ tab.cap     : chr "iris"
#>  $ tab.id      : chr "iris"
#>  $ no_hband    : logi TRUE
#>  $ first_column: logi FALSE
#> 
#>   |                                                                              |..........................................                            |  60%
#>    inline R code fragments
#> 
#>   |                                                                              |.............................................                         |  64%
#> label: unnamed-chunk-8 (with options) 
#> List of 6
#>  $ echo      : logi FALSE
#>  $ fig.cap   : chr "sin function"
#>  $ fig.id    : chr "sinus"
#>  $ fig.width : num 6
#>  $ fig.height: num 4
#>  $ fig.align : chr "center"
#> 

#>   |                                                                              |................................................                      |  68%
#>    inline R code fragments
#> 
#>   |                                                                              |..................................................                    |  72%
#> label: unnamed-chunk-9
#>   |                                                                              |.....................................................                 |  76%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: unnamed-chunk-10
#>   |                                                                              |...........................................................           |  84%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................................................        |  88%
#> label: unnamed-chunk-11 (with options) 
#> List of 3
#>  $ fig.cap      : chr "economics plot"
#>  $ fig.id       : chr "ts-plot"
#>  $ fig.cap.style: chr "Image Caption"
#> 

#>   |                                                                              |................................................................      |  92%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................................................   |  96%
#> label: unnamed-chunk-12 (with options) 
#> List of 2
#>  $ fig.cap: chr "$\\sum_{i=1}^n X_i$"
#>  $ fig.id : chr "dotplot"
#> 

#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: bookdown.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS bookdown.utf8.md --to docx --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output bookdown.docx --highlight-style tango --reference-doc template.docx --lua-filter /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/lua/pagebreak.lua 

#> 
#> Output created: _book/bookdown.docx

docx_file_2 <- file.path(temp_dir, "_book", "bookdown.docx")

if(file.exists(docx_file_2)){
  message("file ", docx_file_2, " has been written.")
}

#> file /var/folders/08/2qdvv0q95wn52xy6mxgj340r0000gn/T//Rtmp1NHdZv/file23bd72804f50/_book/bookdown.docx has been written.