Advanced R Markdown PowerPoint Format

Format for converting from R Markdown to an MS PowerPoint document. rpptx_document2 also supports cross reference based on the syntax of the bookdown package.

The function will allow you to specify the destination of your chunks in the output PowerPoint file. In this case, you must specify the layout and master for the layout you want to use, as well as the ph argument, which will allow you to specify the placeholder to be generated to place the result. Use the officer package to help you choose the identfiers to use.

This function also support Vector graphics output in an editable format (using package rvg). Wrap you R plot commands with function dml to use this graphic capability.

rpptx_document(
  base_format = "rmarkdown::powerpoint_presentation",
  layout = "Title and Content",
  master = "Office Theme",
  tcf = list(),
  ...
)

Arguments

base_format
a scalar character, format to be used as a base document for officedown. default to powerpoint_presentation but can also be powerpoint_presentation2 from bookdown

layout
default slide layout name to use

master
default master layout name where layout is located

tcf
default conditional formatting settings defined by officer::table_conditional_formatting()

...
arguments used by powerpoint_presentation

Value

R Markdown output format to pass to render

Examples

library(rmarkdown)
if(pandoc_available() && require("ggplot2")){
  skeleton <- system.file(package = "officedown",
    "rmarkdown/templates/powerpoint/skeleton/skeleton.Rmd")
  rmd_file <- tempfile(fileext = ".Rmd")
  file.copy(skeleton, to = rmd_file)

  pptx_file_1 <- tempfile(fileext = ".pptx")
  render(rmd_file, output_file = pptx_file_1)
}

#> 
#> 
#> processing file: filea4586bb8ec00.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |...                                                                   |   4%
#>    inline R code fragments
#> 
#>   |                                                                              |.....                                                                 |   7%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |........                                                              |  11%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........                                                            |  15%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ layout: chr "Two Content"
#>  $ ph    : language officer::ph_location_left()
#> 
#>   |                                                                              |.............                                                         |  19%
#>   ordinary text without R code
#> 
#>   |                                                                              |................                                                      |  22%
#> label: unnamed-chunk-2 (with options) 
#> List of 2
#>  $ layout: chr "Two Content"
#>  $ ph    : language officer::ph_location_right()
#> 
#>   |                                                                              |..................                                                    |  26%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................                                                 |  30%
#> label: plot1

#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |..........................                                            |  37%
#> label: unnamed-chunk-3 (with options) 
#> List of 2
#>  $ layout: chr "Title and Content"
#>  $ ph    : language officer::ph_location_type(type = "body")
#> 

#>   |                                                                              |.............................                                         |  41%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................                                       |  44%
#> label: unnamed-chunk-4 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location_fullsize()
#> 

#>   |                                                                              |..................................                                    |  48%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................................                                  |  52%
#> label: unnamed-chunk-5 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location_left()
#> 

#>   |                                                                              |.......................................                               |  56%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........................................                             |  59%
#> label: unnamed-chunk-6 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location_right()
#> 

#>   |                                                                              |............................................                          |  63%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: unnamed-chunk-7 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location_left()
#> 

#>   |                                                                              |.................................................                     |  70%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................................................                  |  74%
#> label: unnamed-chunk-8 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location_right()
#> 

#>   |                                                                              |......................................................                |  78%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........................................................             |  81%
#> label: unnamed-chunk-9 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location(width = 2, height = 2, left = 2, top = 2)
#> 

#>   |                                                                              |............................................................          |  85%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................................................        |  89%
#> label: unnamed-chunk-10 (with options) 
#> List of 1
#>  $ ph: language officer::ph_location(width = 3, height = 3, left = 4, top = 4)
#> 

#>   |                                                                              |.................................................................     |  93%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................................................   |  96%
#> label: unnamed-chunk-11

#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: filea4586bb8ec00.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS filea4586bb8ec00.utf8.md --to pptx --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output /var/folders/08/2qdvv0q95wn52xy6mxgj340r0000gn/T//Rtmp2dza2W/filea458398e2b40.pptx --reference-doc /var/folders/08/2qdvv0q95wn52xy6mxgj340r0000gn/T//Rtmp2dza2W/filea45813b550d9.pptx 

#> 
#> Output created: /var/folders/08/2qdvv0q95wn52xy6mxgj340r0000gn/T//Rtmp2dza2W/filea458398e2b40.pptx